Output just the last http body response.
