+++ integration/tests/assert_base64.out
@@ -1,0 +1,3 @@
+line1
+line2
+line3
