+++ integration/tests/assert_json.out
@@ -1,0 +1,6 @@
+{
+  "success": false,
+  "errors": [{"id":"error1"},{"id":"error2"}],
+  "warnings": [],
+  "duration": 1.5
+}
