+++ integration/tests/assert_match.out
@@ -1,0 +1,4 @@
+{
+  "date1": "2014-01-01",
+  "date2": "x2014-01-01"
+}
