+++ integration/tests/assert_regex.out
@@ -1,0 +1,1 @@
+Hello World!
