+++ integration/tests/assert_xpath.out
@@ -1,0 +1,1 @@
+<data>café</data>
