+++ integration/tests/bytes.out
@@ -1,0 +1,1 @@
+ÿ
