+++ integration/tests/capture_and_assert.out
@@ -1,0 +1,1 @@
+Hello World!
