+++ integration/tests/captures.out
@@ -1,0 +1,1 @@
+{ "a_null": null, "an_object": {"id": "123"}, "a_list": [1,2,3], "an_integer": 1, "a_float": 1.1, "a_bool": true, "a_string": "hello" }
