+++ integration/tests/encoding.out
@@ -1,0 +1,1 @@
+café
