+++ integration/tests/hello.out
@@ -1,0 +1,1 @@
+Hello World!
