+++ integration/tests/output.out
@@ -1,0 +1,1 @@
+Response endpoint2
