+++ integration/tests/utf8.out
@@ -1,0 +1,1 @@
+<data>café</data>
